% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FWERSS.R
\name{MBonf.p.adjust}
\alias{MBonf.p.adjust}
\title{The adjusted p-values for Modified Bonferroni single-step FWER controlling procedure.}
\usage{
MBonf.p.adjust(p, p.set, alpha, make.decision)
}
\arguments{
\item{p}{numeric vector of p-values (possibly with \code{\link[base]{NA}}s). Any other R is coerced by \code{\link[base]{as.numeric}}. Same as in \code{\link[stats]{p.adjust}}.}

\item{p.set}{a list of numeric vectors, where each vector is the vector of all attainable p-values containing the available p-value for the corresponding hypothesis.}

\item{alpha}{significant level used to compare with adjusted p-values to make decisions, the default value is 0.05.}

\item{make.decision}{logical; if  \code{TRUE}, then the output include the decision rules compared adjusted p-values with significant level \eqn{\alpha}}
}
\value{
A numeric vector of the adjusted p-values (of the same length as \code{p}) if  \code{make.decision = FALSE}, or a list including original p-values, adjusted p-values and decision rules if \code{make.decision = TRUE}.
}
\description{
The function for calculating the adjusted p-values based on original available p-values and all attaianble p-values
}
\note{
The attainable p-value refers to the element of domain set of p-value for the corresponding hypothesis. For continuous test statistics, the p-value under true null are uniform distributed in (0,1), thus the p-values are attainable everywhere between 0 and 1. But for discrete test statistics, the p-value can only take finite values bewtween 0 and 1, that is the attainable p-values for discrete case are finite and countable, so we can assign them in a finite list \code{p.set}.
}
\examples{
p <- c(pbinom(1,8,0.5),pbinom(1,5,0.75),pbinom(1,6,0.6))
p.set <-list(pbinom(0:8,8,0.5),pbinom(0:5,5,0.75),pbinom(0:6,6,0.6))
MBonf.p.adjust(p,p.set)
## Compare with the traditional Bonferroni adjustment
p.adjust(p,method = "bonferroni")
## Compare with the Tarone adjustment
Tarone.p.adjust(p,p.set)
}
\references{
Zhu, Y., & Guo, W. (2017).
Familywise error rate controlling procedures for discrete data
\emph{arXiv preprint} arXiv:1711.08147.
}
\seealso{
\code{\link{Tarone.p.adjust}},  \code{\link{MixBonf.p.adjust}},  \code{\link[stats]{p.adjust}}.
}
\author{
Yalin Zhu
}
