% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHTdiscrete.R
\docType{package}
\name{MHTdiscrete}
\alias{MHTdiscrete}
\alias{MHTdiscrete-package}
\title{MHTdiscrete: A package for Multiple Hypotheses Testing for Discrete Data.}
\description{
The MHTdiscrete package provides three categories of important functions:
MBonf.p.adjust, MHolm.p.adjust and MHoch.p.adjust.
}

