% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FWERSS.R
\name{Sidak.p.adjust}
\alias{Sidak.p.adjust}
\title{The adjusted p-values for Sidak single-step FWER controlling procedure.}
\usage{
Sidak.p.adjust(p, alpha, make.decision)
}
\arguments{
\item{p}{numeric vector of p-values (possibly with \code{\link[base]{NA}}s). Any other R is coerced by \code{\link[base]{as.numeric}}. Same as in \code{\link[stats]{p.adjust}}.}

\item{alpha}{significant level used to compare with adjusted p-values to make decisions, the default value is 0.05.}

\item{make.decision}{logical; if  \code{TRUE}, then the output include the decision rules compared adjusted p-values with significant level \eqn{\alpha}}
}
\value{
A numeric vector of the adjusted p-values (of the same length as \code{p}) if  \code{make.decision = FALSE}, or a list including original p-values, adjusted p-values and decision rules if \code{make.decision = TRUE}.
}
\description{
The function for calculating the adjusted p-values based on original available p-values.
}
\examples{
p <- c(pbinom(1,8,0.5),pbinom(1,5,0.75),pbinom(1,6,0.6))
Sidak.p.adjust(p)
}
\seealso{
\code{\link[stats]{p.adjust}}.
}
\author{
Yalin Zhu
}
