% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHTdiscrete.R
\docType{package}
\name{MHTdiscrete}
\alias{MHTdiscrete}
\alias{MHTdiscrete-package}
\title{MHTdiscrete: A package for Multiple Hypotheses Testing for Discrete Data.}
\description{
The MHTdiscrete package provides two categories of important functions for discrete data mutliple hypothese testing:
}
\section{FWER controlling procedures}{

       Single-step: MBonf.p.adjust, Tarone.p.adjust.

       Step-down: MHolm.p.adjust, TH.p.adjust.

       Step-up: MHoch.p.adjust, Roth.p.adjust
       .
}

\section{FDR controlling procedures}{


       Step-down: MBL.p.adjust.

       Step-up: MBH.p.adjust, GTBH.p.adjust, MBY.p.adjust, GTBY.p.adjust.
}
\author{
Yalin Zhu
}

